requirements.txt
ortools
pandas
sqlalchemy
mysql-connector-python
matplotlib
pytest
pytest-cov
python-dotenv